backend/requirements.txt
opencv-python==4.9.0.80
numpy==1.26.4
tensorflow==2.15.0
flask==2.3.3
flask-cors==4.0.0
python-multipart==0.0.6
pillow==10.3.0
facenet-pytorch==2.5.3
insightface==0.7.3
onnxruntime==1.17.0